move client into oauth subpackage
