+++ xl_auth/oauth/client/forms.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-"""Client forms."""
+"""OAuth Client forms."""
 
 from __future__ import absolute_import, division, print_function, unicode_literals
